feat: Use vendor/package namespace convention. Move service to infrastructure.
