+++ .gitignore
@@ -1,4 +1,5 @@
 vendor
 bin
+build
 .idea
 *.swp
